chore: added first framework unit test for parses
