+++ package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "tsc && mocha dist/test --recursive",
     "build": "tsc",
     "prepublish": "tsc"
   },
@@ -21,7 +21,9 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/mocha": "^9.1.1",
     "@types/node": "^17.0.2",
+    "mocha": "^10.0.0",
     "typescript": "4.0.3"
   },
   "dependencies": {
