+++ src/library/index.ts
@@ -1,4 +1,3 @@
-export * from './BaseModel'
 export * from './BaseController'
 export * from './ModelController'
 export * from './ModelAdminController'
